+++ solidity/contracts/proxy/DepositFactory.sol
@@ -67,7 +67,7 @@ contract DepositFactory is CloneFactory, TBTCSystemAuthority{
     ///                        Mints the TDT to the function caller.
     //                         (See `TBTCDepositToken` for more info on TDTs).
     /// @return                True if successful, otherwise revert.
-    function createDeposit (uint256 _lotSizeSatoshis) public payable returns(address) {
+    function createDeposit (uint64 _lotSizeSatoshis) public payable returns(address) {
         address cloneAddress = createClone(masterDepositAddress);
 
         TBTCDepositToken(tbtcDepositToken).mint(msg.sender, uint256(cloneAddress));
